+++ SwiftCharts/Layers/ChartGuideLinesLayer.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 
+// TODO refactor this file
 
 public class ChartGuideLinesIterator {
     
@@ -39,6 +40,40 @@ public class ChartGuideLinesIterator {
         }
     }
 }
+
+
+public class ChartGuideLinesIterator2 {
+    
+    public typealias LineDrawer = (p1: CGPoint, p2: CGPoint) -> ()
+    
+    func iterateLines(layer: ChartCoordsSpaceLayer, axis: ChartGuideLinesLayerAxis, axisValuesX: [ChartAxisValue], axisValuesY: [ChartAxisValue], lineDrawer: LineDrawer) {
+        let originScreenLoc = layer.innerFrame.origin
+        let xScreenLocs = axisValuesX.map({layer.xAxis.screenLocForScalar($0.scalar)})
+        let yScreenLocs = axisValuesY.map({layer.yAxis.screenLocForScalar($0.scalar)})
+        
+        if axis == .X || axis == .XAndY {
+            for xScreenLoc in xScreenLocs {
+                let x1 = xScreenLoc
+                let y1 = originScreenLoc.y
+                let x2 = x1
+                let y2 = originScreenLoc.y + layer.innerFrame.height
+                lineDrawer(p1: CGPointMake(x1, y1), p2: CGPointMake(x2, y2))
+            }
+        }
+        
+        if axis == .Y || axis == .XAndY {
+            for yScreenLoc in yScreenLocs {
+                let x1 = originScreenLoc.x
+                let y1 = yScreenLoc
+                let x2 = originScreenLoc.x + layer.innerFrame.width
+                let y2 = y1
+                lineDrawer(p1: CGPointMake(x1, y1), p2: CGPointMake(x2, y2))
+            }
+        }
+    }
+}
+
+
 
 public struct ChartGuideLinesLayerSettings {
     let linesColor: UIColor
@@ -84,6 +119,74 @@ public class ChartGuideLinesLayer: ChartCoordsSpaceLayer {
     }
 }
 
+public class ChartGuideLinesFilteredLayer: ChartCoordsSpaceLayer {
+    
+    private let settings: ChartGuideLinesDottedLayerSettings
+    private let axisValuesX: [ChartAxisValue]
+    private let axisValuesY: [ChartAxisValue]
+    
+    public init(axisX: ChartAxisLayer, yAxis: ChartAxisLayer, innerFrame: CGRect, settings: ChartGuideLinesDottedLayerSettings, axisValuesX: [ChartAxisValue], axisValuesY: [ChartAxisValue]) {
+        self.settings = settings
+        self.axisValuesX = axisValuesX
+        self.axisValuesY = axisValuesY
+        super.init(xAxis: axisX, yAxis: yAxis, innerFrame: innerFrame)
+    }
+    
+    private func drawGuideline(context: CGContextRef, color: UIColor, width: CGFloat, p1: CGPoint, p2: CGPoint, dotWidth: CGFloat, dotSpacing: CGFloat) {
+        
+        CGContextSetStrokeColorWithColor(context, color.CGColor)
+        CGContextSetLineWidth(context, width)
+        
+        let offset = dotWidth + dotSpacing
+        
+        let offsetY = (p2.y - p1.y)
+        let yConstant = offsetY == 0
+        var limit: CGFloat
+        var start: CGFloat
+        
+        if yConstant { //horizontal line
+            limit = p2.x - p1.x
+            start = p1.x
+            
+        } else { //vertical line
+            limit = p2.y - p1.y
+            start = p1.y
+        }
+        
+        limit += start
+        
+        for tmp in stride(from: start, to: limit, by: offset) {
+            
+            var x1: CGFloat
+            var y1: CGFloat
+            var x2: CGFloat
+            var y2: CGFloat
+            
+            if yConstant { //horizontal line
+                x1 = tmp
+                y1 = p1.y
+                x2 = x1 + dotWidth
+                y2 = y1
+                
+            } else { //vertical line
+                x1 = p1.x
+                y1 = tmp
+                x2 = x1
+                y2 = y1 + dotWidth
+            }
+            
+            CGContextMoveToPoint(context, x1, y1)
+            CGContextAddLineToPoint(context, x2, y2)
+            CGContextStrokePath(context)
+        }
+    }
+    
+    override public func chartViewDrawing(#context: CGContextRef, chart: Chart) {
+        ChartGuideLinesIterator2().iterateLines(self, axis: self.settings.axis, axisValuesX: self.axisValuesX, axisValuesY: self.axisValuesY) {(p1, p2) -> () in
+            self.drawGuideline(context, color: self.settings.linesColor, width: self.settings.linesWidth, p1: p1, p2: p2, dotWidth: self.settings.dotWidth, dotSpacing: self.settings.dotSpacing)
+        }
+    }
+}
 
 public struct ChartGuideLinesDottedLayerSettings {
     let linesColor: UIColor
@@ -165,7 +268,7 @@ public class ChartGuideLinesDottedLayer: ChartCoordsSpaceLayer {
     
     override public func chartViewDrawing(#context: CGContextRef, chart: Chart) {
         ChartGuideLinesIterator().iterateLines(self, axis: self.settings.axis, onlyVisibleX: self.onlyVisibleX, onlyVisibleY: self.onlyVisibleY) {(p1, p2) -> () in
-            self.drawGuideline(context, color: self.settings.linesColor, width: self.settings.linesWidth, p1: p1, p2: p2, dotWidth: 2, dotSpacing: 2)
+            self.drawGuideline(context, color: self.settings.linesColor, width: self.settings.linesWidth, p1: p1, p2: p2, dotWidth: self.settings.dotWidth, dotSpacing: self.settings.dotSpacing)
         }
     }
 }